refactor: move surcharges files from `no-meteor` to `node-app`
