+++ imports/node-app/core/util/buildContext.test.js
@@ -1,4 +1,4 @@
-import mockContext from "/imports/test-utils/helpers/mockContext";
+import mockContext from "/imports/test-utils/helpers/mockContext"; // TODO: EK - update to come from new packages
 import buildContext from "./buildContext";
 
 const fakeUser = {
