+++ imports/node-app/plugins/surcharges/index.js
@@ -1,9 +1,9 @@
-import getSurcharges from "./getSurcharges";
-import mutations from "./mutations";
-import queries from "./queries";
-import resolvers from "./resolvers";
-import schemas from "./schemas";
-import setSurchargesOnCart from "./util/setSurchargesOnCart";
+import getSurcharges from "./getSurcharges.js";
+import mutations from "./mutations/index.js";
+import queries from "./queries/index.js";
+import resolvers from "./resolvers/index.js";
+import schemas from "./schemas/index.js";
+import setSurchargesOnCart from "./util/setSurchargesOnCart.js";
 
 /**
  * @summary Import and call this function to add this plugin to your API.
