+++ imports/node-app/plugins/surcharges/mutations/createSurcharge.js
@@ -1,26 +1,25 @@
 import SimpleSchema from "simpl-schema";
+import Random from "@reactioncommerce/random";
 import ReactionError from "@reactioncommerce/reaction-error";
-import surchargeSchema from "../util/surchargeSchema";
+import surchargeSchema from "../util/surchargeSchema.js";
 
 const inputSchema = new SimpleSchema({
-  shopId: String,
   surcharge: surchargeSchema,
-  surchargeId: String
+  shopId: String
 });
 
-
 /**
- * @method updateSurchargeMutation
- * @summary updates a surcharge
+ * @method createSurchargeMutation
+ * @summary Creates a surcharge
  * @param {Object} context - an object containing the per-request state
  * @param {Object} input - Input (see SimpleSchema)
- * @returns {Promise<Object>} An object with a `surcharge` property containing the updated surcharge
+ * @returns {Promise<Object>} An object with a `surcharge` property containing the created surcharge
  */
-export default async function updateSurchargeMutation(context, input) {
+export default async function createSurchargeMutation(context, input) {
   const cleanedInput = inputSchema.clean(input); // add default values and such
   inputSchema.validate(cleanedInput);
 
-  const { surcharge, surchargeId, shopId } = cleanedInput;
+  const { surcharge, shopId } = cleanedInput;
   const { collections, userHasPermission } = context;
   const { Surcharges } = collections;
 
@@ -28,16 +27,14 @@ export default async function updateSurchargeMutation(context, input) {
     throw new ReactionError("access-denied", "Access Denied");
   }
 
-  const { matchedCount } = await Surcharges.updateOne({
-    _id: surchargeId,
-    shopId
-  }, {
-    $set: {
-      updatedAt: new Date(),
-      ...surcharge
-    }
+  surcharge._id = Random.id();
+
+  const { insertedCount } = await Surcharges.insertOne({
+    shopId,
+    createdAt: new Date(),
+    ...surcharge
   });
-  if (matchedCount === 0) throw new ReactionError("not-found", "Not found");
+  if (insertedCount === 0) throw new ReactionError("server-error", "Unable to create surcharge");
 
   return { surcharge };
 }
