+++ imports/node-app/plugins/surcharges/queries/index.js
@@ -1,0 +1,7 @@
+import surchargeById from "./surchargeById.js";
+import surcharges from "./surcharges.js";
+
+export default {
+  surchargeById,
+  surcharges
+};
