+++ imports/node-app/plugins/surcharges/resolvers/Mutation/createSurcharge.js
@@ -1,6 +1,6 @@
 import { decodeFulfillmentMethodOpaqueId } from "@reactioncommerce/reaction-graphql-xforms/fulfillment";
 import { decodeShopOpaqueId } from "@reactioncommerce/reaction-graphql-xforms/shop";
-import createSurchargeMutation from "../../mutations/createSurcharge";
+import createSurchargeMutation from "../../mutations/createSurcharge.js";
 
 /**
  * @name Mutation/createSurcharge
