+++ imports/node-app/plugins/surcharges/resolvers/Mutation/index.js
@@ -1,0 +1,9 @@
+import createSurcharge from "./createSurcharge.js";
+import deleteSurcharge from "./deleteSurcharge.js";
+import updateSurcharge from "./updateSurcharge.js";
+
+export default {
+  createSurcharge,
+  deleteSurcharge,
+  updateSurcharge
+};
