+++ imports/node-app/plugins/surcharges/resolvers/Surcharge/index.js
@@ -1,8 +1,8 @@
 import { encodeFulfillmentMethodOpaqueId } from "@reactioncommerce/reaction-graphql-xforms/fulfillment";
 import { encodeShopOpaqueId } from "@reactioncommerce/reaction-graphql-xforms/shop";
-import { encodeSurchargeOpaqueId } from "../../xforms/surcharge";
-import xformSurchargeAmount from "../../xforms/xformSurchargeAmount";
-import getSurchargeMessageForLanguage from "../../util/getSurchargeMessageForLanguage";
+import { encodeSurchargeOpaqueId } from "../../xforms/surcharge.js";
+import xformSurchargeAmount from "../../xforms/xformSurchargeAmount.js";
+import getSurchargeMessageForLanguage from "../../util/getSurchargeMessageForLanguage.js";
 
 export default {
   _id: (node) => encodeSurchargeOpaqueId(node._id),
