--- imports/plugins/included/surcharges/server/no-meteor/queries/index.js
@@ -1,7 +1,0 @@
-import surchargeById from "./surchargeById";
-import surcharges from "./surcharges";
-
-export default {
-  surchargeById,
-  surcharges
-};
