--- imports/plugins/included/surcharges/server/no-meteor/resolvers/Mutation/index.js
@@ -1,9 +1,0 @@
-import createSurcharge from "./createSurcharge";
-import deleteSurcharge from "./deleteSurcharge";
-import updateSurcharge from "./updateSurcharge";
-
-export default {
-  createSurcharge,
-  deleteSurcharge,
-  updateSurcharge
-};
